+++ PCACalcula.XUnitTest/tests/Utils/Utilidades.cs
@@ -7,7 +7,7 @@ namespace PCACalcula.XUnitTest.tests.Utils
     public static class Utilidades
     {
         public static float CalcularResultadoEsperado(decimal valorInicial, int meses)
-           => ((float)((double)valorInicial * Math.Pow(1.01, meses))).ToTwoPlaces();
+           => (((double)valorInicial * Math.Pow(1.01, meses))).ToTwoPlaces();
 
         public static Notification CreateNotification(string propriedade, string mensagem)
             => new Notification(propriedade, mensagem);
